requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
fpdf
google-cloud-storage
google-auth
google-auth-oauthlib 